Completed the Etch A Sketch project
I used javascript to add most htm components and update the board live. Also used css to help style the container and make the box fill the whole page.
I also made it so the divs fit no mater the amount per row and column but I made a max of 100.